+++ examples/hcaptcha_request_proxy.py
@@ -1,23 +1,23 @@
 from six.moves.urllib import parse
+
+import re
 import requests
 from os import environ
-import re
-from random import choice
 
-from python_anticaptcha import AnticaptchaClient, FunCaptchaTask
+from python_anticaptcha import AnticaptchaClient, HCaptchaTask
 
 api_key = environ["KEY"]
-site_key_pattern = 'public_key: "(.+?)",'
-url = "https://client-demo.arkoselabs.com/solo-animals"
+proxy_url = environ["PROXY_URL"]  # eg. socks5://user:password/123.123.123.123:8888/
+site_key_pattern = 'data-sitekey="(.+?)"'
+url = "http://hcaptcha.jawne.info.pl/"
 client = AnticaptchaClient(api_key)
 session = requests.Session()
+EXPECTED_RESULT = "Your request have submitted successfully."
 
 UA = (
     "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_10_1) AppleWebKit/537.36 "
     "(KHTML, like Gecko) Chrome/39.0.2171.95 Safari/537.36"
 )
-session.headers = {"User-Agent": UA}
-proxy_url = environ["PROXY_URL"]
 
 
 def parse_url(url):
@@ -36,19 +36,29 @@ def get_form_html():
 
 
 def get_token(form_html):
-    proxy = parse_url(proxy_url)
-
     site_key = re.search(site_key_pattern, form_html).group(1)
-    task = FunCaptchaTask(url, site_key, user_agent=UA, **proxy)
+    proxy = parse_url(proxy_url)
+    task = HCaptchaTask(
+        website_url=url,
+        website_key=site_key,
+        user_agent=UA,
+        cookies="test=test",
+        **proxy
+    )
     job = client.createTask(task)
-    job.join(maximum_time=10 ** 4)
-    return job.get_token_response()
+    job.join()
+    return job.get_solution_response()
+
+
+def form_submit(token):
+    return requests.post(url, data={"g-recaptcha-response": token}).text
 
 
 def process():
     html = get_form_html()
-    return get_token(html)
+    token = get_token(html)
+    return form_submit(token)
 
 
 if __name__ == "__main__":
-    print('Solved!' in process())
+    assert EXPECTED_RESULT in process()