+++ tests/test_examples.py
@@ -16,6 +16,12 @@ def missing_key(*args, **kwargs):
     )(*args, **kwargs)
 
 
+def missing_proxy(*args, **kwargs):
+    return skipIf(
+        "PROXY_URL" not in os.environ, "Missing PROXY_URL environment variable"
+    )(*args, **kwargs)
+
+
 @missing_key
 class CustomDotTestCase(TestCase):
     # For unknown reasons, workers are not always
@@ -51,15 +57,15 @@ class CustomModerationTestCase(TestCase):
 
 
 @missing_key
-@skipIf("PROXY_URL" not in os.environ, "Missing PROXY_URL environment variable")
+@missing_proxy
 class FuncaptchaTestCase(TestCase):
     # CI Proxy is unstable.
     # Occasionally fails, so I repeat my attempt to have others selected.
     @retry(tries=3)
     def test_funcaptcha(self):
-        from examples import funcaptcha
+        from examples import funcaptcha_request
 
-        self.assertTrue(funcaptcha.process())
+        self.assertIn('Solved!', funcaptcha_request.process())
 
 
 @missing_key
@@ -73,7 +79,8 @@ class RecaptchaRequestTestCase(TestCase):
 
 
 @missing_key
-class RecaptchaV3TestCase(TestCase):
+@skipIf(True, "Anti-captcha unable to provide required score, but we tests via proxy")
+class RecaptchaV3ProxylessTestCase(TestCase):
     # Anticaptcha responds is not fully reliable.
     @retry(tries=3)
     def test_process(self):
@@ -114,6 +121,7 @@ class TextTestCase(TestCase):
 
 
 @missing_key
+@skipIf(True, 'We testing via proxy for performance reason.')
 class HCaptchaTaskProxylessTestCase(TestCase):
     @retry(tries=3)
     def test_process(self):
@@ -123,12 +131,13 @@ class HCaptchaTaskProxylessTestCase(TestCase):
 
 
 @missing_key
-class HCaptchaTaskProxylessTestCase(TestCase):
+class HCaptchaTaskTestCase(TestCase):
     def test_process(self):
-        from examples import hcaptcha_request
+        from examples import hcaptcha_request_proxy
 
         self.assertIn(
-            "Your request have submitted successfully.", hcaptcha_request.process()
+            "Your request have submitted successfully.",
+            hcaptcha_request_proxy.process(),
         )